+++ scripts/utils/file.ts
@@ -3,7 +3,6 @@ import Package from '../../package.json'
 import { mkdir } from 'fs/promises'
 import fs from 'node:fs/promises'
 import { ChainId, FungibleToken } from '../type'
-import { generateTokenList } from '../../src/helpers/generate-token-list'
 import { EthereumAddress } from 'wallet.ts'
 
 // @ts-ignore
@@ -53,4 +52,40 @@ function generate(tokens: FungibleToken[]) {
   )
 
   return JSON.stringify(tokenList)
+}
+
+/**
+ * generateTokenList.
+ *
+ * @param {Array<object>} specList
+ * @param {Object} extraInfo
+ */
+function generateTokenList(tokens: FungibleToken[], extraInfo: Record<string, string | string[]>) {
+  return {
+    name: 'Mask',
+    logoURI:
+      'https://raw.githubusercontent.com/DimensionDev/Maskbook-Website/master/img/MB--CircleCanvas--WhiteOverBlue.svg',
+    keywords: [
+      'browser extension',
+      'web3',
+      'peer to peer',
+      'encryption',
+      'cryptography',
+      'gundb',
+      'privacy protection',
+      'ownyourdata',
+      'social network',
+      'blockchain',
+      'crypto',
+      'dweb',
+    ],
+    ...extraInfo,
+    timestamp: new Date().toISOString(),
+    version: {
+      major: Number.parseInt(Package.version.split('.')[0]),
+      minor: Number.parseInt(Package.version.split('.')[1]),
+      patch: Number.parseInt(Package.version.split('.')[2]),
+    },
+    tokens,
+  }
 }
